read csv file if present

+++ sic-pred/sic_climatology.py
@@ -4,6 +4,7 @@ import datetime as dt
 import numpy as np
 from matplotlib import pyplot as plt
 from matplotlib import dates as mdates
+import pandas as pd
 
 from sic_pred_base import SicPredClimatology
 
@@ -34,11 +35,15 @@ def run():
     df_all = dict()
     for num_years in range(1, _MAX_YEARS + 1):
         fc = SicPredClimatology(_LATEST_YEAR, num_years)
-        df = fc.comp_all_errors(_START, _END)
         outfile = f'out/climatology-years{num_years}.csv'
-        os.makedirs(os.path.dirname(outfile), exist_ok=True)
-        print(f'Saving {outfile}')
-        df.set_index('Date').to_csv(outfile)
+        if os.path.exists(outfile):
+            print(f'Reading {outfile}')
+            df = pd.read_csv(outfile)
+        else:
+            df = fc.comp_all_errors(_START, _END)
+            os.makedirs(os.path.dirname(outfile), exist_ok=True)
+            print(f'Saving {outfile}')
+            df.set_index('Date').to_csv(outfile)
         df_all[num_years] = df
     make_plots(df_all, 'out/climatology.png')
